requirements.txt
pymcubes
setuptools
lpips
scipy
tqdm
rich
imageio
imageio-ffmpeg
opencv-python
trimesh
plyfile
scikit-image
scikit-learn
icecream
matplotlib
ipympl
jupyterlab
ipython
wandb
chardet
hydra-core
fake-bpy-module-3.2
autopep8
ninja
einops